Move all code for printing linear model to seperate class
