+++ src/rtctools/optimization/optimization_problem.py
@@ -44,8 +44,8 @@ class OptimizationProblem(DataStoreAccessor, metaclass=ABCMeta):
     """
     Base class for all optimization problems.
     """
-    store_intermediate_lp_info = False
     store_linear_problem = False
+    manual_expansion = False
     lam_tol = 1.5
     _debug_check_level = DebugLevel.MEDIUM
     _debug_check_options = {}
@@ -95,7 +95,7 @@ class OptimizationProblem(DataStoreAccessor, metaclass=ABCMeta):
         options.update(self.solver_options())  # Create a copy
 
         logger.debug("Creating solver")
-        if options.pop("expand", False) or self.store_intermediate_lp_info:
+        if options.pop("expand", False) or self.manual_expansion:
             # NOTE: CasADi only supports the "expand" option for nlpsol. To
             # also be able to expand with e.g. qpsol, we do the expansion
             # ourselves here.
@@ -108,18 +108,6 @@ class OptimizationProblem(DataStoreAccessor, metaclass=ABCMeta):
             nlp["f"] = f_sx
             nlp["g"] = g_sx
             nlp["x"] = X_sx
-
-            expand_f_g = ca.Function("f_g", [nlp["x"]], [nlp["f"], nlp["g"]]).expand()
-            casadi_equations = {}
-            casadi_equations["indices"] = self._CollocatedIntegratedOptimizationProblem__indices
-            casadi_equations["func"] = expand_f_g
-            casadi_equations["other"] = (lbx, ubx, lbg, ubg, x0)
-            in_var = ca.SX.sym("X", expand_f_g.nnz_in())
-            bf = ca.Function("bf", [in_var], [expand_f_g(in_var)[1]])
-            b = bf(0)
-            b = ca.sparsify(b)
-            b = np.array(b)[:, 0]
-            converted_constraints, constraints_original, variable_names = casadi_to_lp(casadi_equations)
 
         # Debug check for non-linearity in constraints
         self.__debug_check_linearity_constraints(nlp)
@@ -161,6 +149,7 @@ class OptimizationProblem(DataStoreAccessor, metaclass=ABCMeta):
         logger.info("Calling solver")
 
         results = solver(x0=x0, lbx=lbx, ubx=ubx, lbg=ca.veccat(*lbg), ubg=ca.veccat(*ubg))
+        self.store_results(results, nlp, lbx, ubx, lbg, ubg, x0)
 
         # Extract relevant stats
         self.__objective_value = float(results["f"])
@@ -196,99 +185,6 @@ class OptimizationProblem(DataStoreAccessor, metaclass=ABCMeta):
             except (AttributeError, ValueError):
                 logger.log(log_level, "Solver succeeded with status {} ({}).".format(
                     return_status, wall_clock_time))
-
-        if self.store_intermediate_lp_info:
-            # Evaluate the constraints wrt to the optimized solution
-            x_optimized = np.array(results["x"]).ravel()
-            expand_f_g = ca.Function("f_g", [nlp["x"]], [nlp["f"], nlp["g"]]).expand()
-            X_sx = ca.SX.sym("X", *nlp["x"].shape)
-            f_sx, g_sx = expand_f_g(X_sx)
-            eval_g = ca.Function("g_eval", [X_sx], [g_sx]).expand()
-            evaluated_g = [x[0] for x in np.array(eval_g(x_optimized))]
-            lam_g = [x[0] for x in np.array(results["lam_g"])]
-            lam_x = [x[0] for x in np.array(results["lam_x"])]
-
-            def extract_var_name_timestep(variable):
-                """Split the variable name into its original name and its timestep"""
-                var_name, _, timestep_str = variable.partition("__")
-                return var_name, int(timestep_str)
-
-            def add_to_dict(new_dict, var_name, timestep, sign="+"):
-                """Add variable to dict grouped by variable names"""
-                if var_name not in new_dict:
-                    new_dict[var_name] = {"timesteps": [timestep], "effect_direction": sign}
-                else:
-                    new_dict[var_name]["timesteps"].append(timestep)
-                return new_dict
-
-            def convert_to_dict_per_var(constrain_list):
-                """Convert list of ungrouped variables to a dict per variable name,
-                with as values the time-indices where the variable was active"""
-                new_dict = {}
-                for constrain in constrain_list:
-                    if isinstance(constrain, list):
-                        for i, variable in enumerate(constrain[2::3]):
-                            var_name, timestep = extract_var_name_timestep(variable)
-                            add_to_dict(new_dict, var_name, timestep, constrain[i * 3])
-                    else:
-                        var_name, timestep = extract_var_name_timestep(constrain)
-                        add_to_dict(new_dict, var_name, timestep)
-                # Sort values and remove duplicates
-                for var_name in new_dict:
-                    new_dict[var_name]["timesteps"] = sorted(set(new_dict[var_name]["timesteps"]))
-                return new_dict
-
-            def check_lambda_exceedence(lam, tol):
-                return [x > tol for x in lam], [x < -tol for x in lam]
-
-            def find_variable_hits(exceedance_list, lowers, uppers, variable_names, variable_values, lam):
-                variable_hits = []
-                for i, hit in enumerate(exceedance_list):
-                    if hit:
-                        logger.debug("Bound for variable {}={} was hit! Lam={}".format(
-                            variable_names[i], variable_values[i], lam))
-                        logger.debug("{} < {} < {}".format(lowers[i], variable_values[i], uppers[i]))
-                        variable_hits.append(variable_names[i])
-                return variable_hits
-
-            # Upper and lower bounds
-            lam_x_larger_than_zero, lam_x_smaller_than_zero = check_lambda_exceedence(lam_x, self.lam_tol)
-            self.upper_bound_variable_hits = find_variable_hits(
-                lam_x_larger_than_zero, lbx, ubx, variable_names, x_optimized, lam_x)
-            self.lower_bound_variable_hits = find_variable_hits(
-                lam_x_smaller_than_zero, lbx, ubx, variable_names, x_optimized, lam_x)
-            self.upper_bound_dict = convert_to_dict_per_var(self.upper_bound_variable_hits)
-            self.lower_bound_dict = convert_to_dict_per_var(self.lower_bound_variable_hits)
-
-            # Upper and lower constraints
-            lam_g_larger_than_zero, lam_g_smaller_than_zero = check_lambda_exceedence(lam_g, self.lam_tol)
-            self.upper_constraint_variable_hits = find_variable_hits(
-                lam_g_larger_than_zero, lbg, ubg, constraints_original, evaluated_g, lam_g)
-            self.lower_constraint_variable_hits = find_variable_hits(
-                lam_g_smaller_than_zero, lbg, ubg, constraints_original, evaluated_g, lam_g)
-            self.upper_constraint_dict = convert_to_dict_per_var(self.upper_constraint_variable_hits)
-            self.lower_constraint_dict = convert_to_dict_per_var(self.lower_constraint_variable_hits)
-
-            # Also save list of active constraints
-            self.active_upper_constraints = [
-                constraint for i, constraint in enumerate(converted_constraints) if lam_g_larger_than_zero[i]
-                ]
-            self.active_lower_constraints = [
-                constraint for i, constraint in enumerate(converted_constraints) if lam_g_smaller_than_zero[i]
-                ]
-
-        # --> Unused function, but could be useful for refactoring convert_to_dict_per_var
-        # get_variables_in_constraints(self.upper_constraint_variable_hits)
-        # def get_variables_in_constraints(constraints):
-        #     variables_in_constraints = []
-        #     for constraint in constraints:
-        #         variables_in_constraints.append([])
-        #         for var_i in range(int(len(constraint)/3)):
-        #             var_sign  = constraint[  var_i*3]
-        #             var_value = constraint[1+var_i*3]
-        #             var_name  = constraint[2+var_i*3]
-        #             variables_in_constraints[-1].append(var_name)
-        #     return variables_in_constraints
 
         # Do any postprocessing
         if postprocessing:
@@ -823,6 +719,12 @@ class OptimizationProblem(DataStoreAccessor, metaclass=ABCMeta):
     def post(self) -> None:
         """
         Postprocessing logic is performed here.
+        """
+        pass
+
+    def store_results(self, results, nlp, lbx, ubx, lbg, ubg, x0) -> None:
+        """
+        Called after solver is finished.
         """
         pass
